Cleanup: core & create packages (#90)
* Cleanup: core & create repos

* Fix import
