+++ packages/govern-create/lib/utils.js
@@ -1,0 +1,6 @@
+const format = ({ address }, name) =>
+  `- ${name}: https://rinkeby.etherscan.io/address/${address}`
+
+const print = (contract, name) => console.log(format(contract, name))
+
+module.exports = { format, print }
