+++ packages/govern-create/tasks/accounts.js
@@ -1,0 +1,7 @@
+module.exports = async () => {
+  const accounts = await ethers.getSigners()
+
+  for (const account of accounts) {
+    console.log(await account.getAddress())
+  }
+}
